Move enum ContactType to Command class
