+++ src/test/java/seedu/address/logic/commands/EditCommandTest.java
@@ -41,7 +41,7 @@ public class EditCommandTest {
     public void execute_allFieldsSpecifiedUnfilteredList_success() {
         Contact editedContact = new PersonBuilder().build();
         EditCommand.EditContactDescriptor descriptor = new EditPersonDescriptorBuilder(editedContact).build();
-        EditCommand editCommand = new EditCommand(INDEX_FIRST_PERSON, descriptor, EditCommand.ContactType.CLIENT);
+        EditCommand editCommand = new EditCommand(INDEX_FIRST_PERSON, descriptor, Command.ContactType.CLIENT);
 
         String expectedMessage = String.format(EditCommand.MESSAGE_EDIT_PERSON_SUCCESS, editedContact);
 
@@ -63,7 +63,7 @@ public class EditCommandTest {
 
         EditCommand.EditContactDescriptor descriptor = new EditPersonDescriptorBuilder().withName(VALID_NAME_BOB)
                 .withPhone(VALID_PHONE_BOB).withTags(VALID_TAG_HUSBAND).build();
-        EditCommand editCommand = new EditCommand(indexLastPerson, descriptor, EditCommand.ContactType.CLIENT);
+        EditCommand editCommand = new EditCommand(indexLastPerson, descriptor, Command.ContactType.CLIENT);
 
         String expectedMessage = String.format(EditCommand.MESSAGE_EDIT_PERSON_SUCCESS, editedContact);
 
@@ -77,7 +77,7 @@ public class EditCommandTest {
     @Test
     public void execute_noFieldSpecifiedUnfilteredList_success() {
         EditCommand editCommand = new EditCommand(INDEX_FIRST_PERSON, new EditContactDescriptor(),
-                EditCommand.ContactType.CLIENT);
+                Command.ContactType.CLIENT);
         Contact editedContact = model.getFilteredContactList().get(INDEX_FIRST_PERSON.getZeroBased());
 
         String expectedMessage = String.format(EditCommand.MESSAGE_EDIT_PERSON_SUCCESS, editedContact);
@@ -96,7 +96,7 @@ public class EditCommandTest {
         Contact editedContact = new PersonBuilder(contactInFilteredList).withName(VALID_NAME_BOB).build();
         EditCommand editCommand = new EditCommand(INDEX_FIRST_PERSON,
                 new EditPersonDescriptorBuilder().withName(VALID_NAME_BOB).build(),
-                EditCommand.ContactType.CLIENT);
+                Command.ContactType.CLIENT);
 
         String expectedMessage = String.format(EditCommand.MESSAGE_EDIT_PERSON_SUCCESS, editedContact);
 
@@ -111,7 +111,7 @@ public class EditCommandTest {
     public void execute_duplicatePersonUnfilteredList_failure() {
         Contact firstContact = model.getFilteredContactList().get(INDEX_FIRST_PERSON.getZeroBased());
         EditCommand.EditContactDescriptor descriptor = new EditPersonDescriptorBuilder(firstContact).build();
-        EditCommand editCommand = new EditCommand(INDEX_SECOND_PERSON, descriptor, EditCommand.ContactType.CLIENT);
+        EditCommand editCommand = new EditCommand(INDEX_SECOND_PERSON, descriptor, Command.ContactType.CLIENT);
 
         assertCommandFailure(editCommand, model, commandHistory, EditCommand.MESSAGE_DUPLICATE_PERSON);
     }
@@ -123,7 +123,7 @@ public class EditCommandTest {
         // edit person in filtered list into a duplicate in address book
         Contact contactInList = model.getAddressBook().getContactList().get(INDEX_SECOND_PERSON.getZeroBased());
         EditCommand editCommand = new EditCommand(INDEX_FIRST_PERSON,
-                new EditPersonDescriptorBuilder(contactInList).build(), EditCommand.ContactType.CLIENT);
+                new EditPersonDescriptorBuilder(contactInList).build(), Command.ContactType.CLIENT);
 
         assertCommandFailure(editCommand, model, commandHistory, EditCommand.MESSAGE_DUPLICATE_PERSON);
     }
@@ -132,7 +132,7 @@ public class EditCommandTest {
     public void execute_invalidPersonIndexUnfilteredList_failure() {
         Index outOfBoundIndex = Index.fromOneBased(model.getFilteredContactList().size() + 1);
         EditContactDescriptor descriptor = new EditPersonDescriptorBuilder().withName(VALID_NAME_BOB).build();
-        EditCommand editCommand = new EditCommand(outOfBoundIndex, descriptor, EditCommand.ContactType.CLIENT);
+        EditCommand editCommand = new EditCommand(outOfBoundIndex, descriptor, Command.ContactType.CLIENT);
 
         assertCommandFailure(editCommand, model, commandHistory, Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
     }
@@ -149,7 +149,7 @@ public class EditCommandTest {
         assertTrue(outOfBoundIndex.getZeroBased() < model.getAddressBook().getContactList().size());
 
         EditCommand editCommand = new EditCommand(outOfBoundIndex,
-                new EditPersonDescriptorBuilder().withName(VALID_NAME_BOB).build(), EditCommand.ContactType.CLIENT);
+                new EditPersonDescriptorBuilder().withName(VALID_NAME_BOB).build(), Command.ContactType.CLIENT);
 
         assertCommandFailure(editCommand, model, commandHistory, Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
     }
@@ -159,7 +159,7 @@ public class EditCommandTest {
         Contact editedContact = new PersonBuilder().build();
         Contact contactToEdit = model.getFilteredContactList().get(INDEX_FIRST_PERSON.getZeroBased());
         EditContactDescriptor descriptor = new EditPersonDescriptorBuilder(editedContact).build();
-        EditCommand editCommand = new EditCommand(INDEX_FIRST_PERSON, descriptor, EditCommand.ContactType.CLIENT);
+        EditCommand editCommand = new EditCommand(INDEX_FIRST_PERSON, descriptor, Command.ContactType.CLIENT);
         Model expectedModel = new ModelManager(new AddressBook(model.getAddressBook()), new UserPrefs());
         expectedModel.updateContact(contactToEdit, editedContact);
         expectedModel.commitAddressBook();
@@ -180,7 +180,7 @@ public class EditCommandTest {
     public void executeUndoRedo_invalidIndexUnfilteredList_failure() {
         Index outOfBoundIndex = Index.fromOneBased(model.getFilteredContactList().size() + 1);
         EditContactDescriptor descriptor = new EditPersonDescriptorBuilder().withName(VALID_NAME_BOB).build();
-        EditCommand editCommand = new EditCommand(outOfBoundIndex, descriptor, EditCommand.ContactType.CLIENT);
+        EditCommand editCommand = new EditCommand(outOfBoundIndex, descriptor, Command.ContactType.CLIENT);
 
         // execution failed -> address book state not added into model
         assertCommandFailure(editCommand, model, commandHistory, Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
@@ -201,7 +201,7 @@ public class EditCommandTest {
     public void executeUndoRedo_validIndexFilteredList_samePersonEdited() throws Exception {
         Contact editedContact = new PersonBuilder().build();
         EditContactDescriptor descriptor = new EditPersonDescriptorBuilder(editedContact).build();
-        EditCommand editCommand = new EditCommand(INDEX_FIRST_PERSON, descriptor, EditCommand.ContactType.CLIENT);
+        EditCommand editCommand = new EditCommand(INDEX_FIRST_PERSON, descriptor, Command.ContactType.CLIENT);
         Model expectedModel = new ModelManager(new AddressBook(model.getAddressBook()), new UserPrefs());
 
         showPersonAtIndex(model, INDEX_SECOND_PERSON);
@@ -225,12 +225,12 @@ public class EditCommandTest {
     @Test
     public void equals() {
         final EditCommand standardCommand = new EditCommand(INDEX_FIRST_PERSON, DESC_AMY,
-                EditCommand.ContactType.CLIENT);
+                Command.ContactType.CLIENT);
 
         // same values -> returns true
         EditContactDescriptor copyDescriptor = new EditCommand.EditContactDescriptor(DESC_AMY);
         EditCommand commandWithSameValues = new EditCommand(INDEX_FIRST_PERSON, copyDescriptor,
-                EditCommand.ContactType.CLIENT);
+                Command.ContactType.CLIENT);
         assertTrue(standardCommand.equals(commandWithSameValues));
 
         // same object -> returns true
@@ -244,11 +244,11 @@ public class EditCommandTest {
 
         // different index -> returns false
         assertFalse(standardCommand.equals(new EditCommand(INDEX_SECOND_PERSON, DESC_AMY,
-                EditCommand.ContactType.CLIENT)));
+                Command.ContactType.CLIENT)));
 
         // different descriptor -> returns false
         assertFalse(standardCommand.equals(new EditCommand(INDEX_FIRST_PERSON, DESC_BOB,
-                EditCommand.ContactType.CLIENT)));
+                Command.ContactType.CLIENT)));
     }
 
 }
