+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -16,6 +16,7 @@ import org.junit.rules.ExpectedException;
 
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.commands.ClearCommand;
+import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.DeleteCommand;
 import seedu.address.logic.commands.EditCommand;
 import seedu.address.logic.commands.ExitCommand;
@@ -65,7 +66,7 @@ public class AddressBookParserTest {
         EditCommand.EditContactDescriptor descriptor = new EditPersonDescriptorBuilder(contact).build();
         EditCommand command = (EditCommand) parser.parseCommand(EditCommand.COMMAND_WORD + " "
                 + INDEX_FIRST_PERSON.getOneBased() + " " + PersonUtil.getEditPersonDescriptorDetails(descriptor));
-        assertEquals(new EditCommand(INDEX_FIRST_PERSON, descriptor, EditCommand.ContactType.CLIENT), command);
+        assertEquals(new EditCommand(INDEX_FIRST_PERSON, descriptor, Command.ContactType.CLIENT), command);
     }
 
     @Test
