+++ src/test/java/seedu/address/logic/parser/EditCommandParserTest.java
@@ -34,6 +34,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_THIRD_PERSON;
 import org.junit.Test;
 
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.EditCommand;
 import seedu.address.logic.commands.EditCommand.EditContactDescriptor;
 import seedu.address.model.contact.Address;
@@ -114,7 +115,7 @@ public class EditCommandParserTest {
         EditContactDescriptor descriptor = new EditPersonDescriptorBuilder().withName(VALID_NAME_AMY)
                 .withPhone(VALID_PHONE_BOB).withEmail(VALID_EMAIL_AMY).withAddress(VALID_ADDRESS_AMY)
                 .withTags(VALID_TAG_HUSBAND, VALID_TAG_FRIEND).build();
-        EditCommand expectedCommand = new EditCommand(targetIndex, descriptor, EditCommand.ContactType.CLIENT);
+        EditCommand expectedCommand = new EditCommand(targetIndex, descriptor, Command.ContactType.CLIENT);
 
         assertParseSuccess(parser, userInput, expectedCommand);
     }
@@ -126,7 +127,7 @@ public class EditCommandParserTest {
 
         EditContactDescriptor descriptor = new EditPersonDescriptorBuilder().withPhone(VALID_PHONE_BOB)
                 .withEmail(VALID_EMAIL_AMY).build();
-        EditCommand expectedCommand = new EditCommand(targetIndex, descriptor, EditCommand.ContactType.CLIENT);
+        EditCommand expectedCommand = new EditCommand(targetIndex, descriptor, Command.ContactType.CLIENT);
 
         assertParseSuccess(parser, userInput, expectedCommand);
     }
@@ -138,31 +139,31 @@ public class EditCommandParserTest {
         String userInput = targetIndex.getOneBased() + NAME_DESC_AMY;
         EditCommand.EditContactDescriptor descriptor = new EditPersonDescriptorBuilder()
                 .withName(VALID_NAME_AMY).build();
-        EditCommand expectedCommand = new EditCommand(targetIndex, descriptor, EditCommand.ContactType.CLIENT);
+        EditCommand expectedCommand = new EditCommand(targetIndex, descriptor, Command.ContactType.CLIENT);
         assertParseSuccess(parser, userInput, expectedCommand);
 
         // phone
         userInput = targetIndex.getOneBased() + PHONE_DESC_AMY;
         descriptor = new EditPersonDescriptorBuilder().withPhone(VALID_PHONE_AMY).build();
-        expectedCommand = new EditCommand(targetIndex, descriptor, EditCommand.ContactType.CLIENT);
+        expectedCommand = new EditCommand(targetIndex, descriptor, Command.ContactType.CLIENT);
         assertParseSuccess(parser, userInput, expectedCommand);
 
         // email
         userInput = targetIndex.getOneBased() + EMAIL_DESC_AMY;
         descriptor = new EditPersonDescriptorBuilder().withEmail(VALID_EMAIL_AMY).build();
-        expectedCommand = new EditCommand(targetIndex, descriptor, EditCommand.ContactType.CLIENT);
+        expectedCommand = new EditCommand(targetIndex, descriptor, Command.ContactType.CLIENT);
         assertParseSuccess(parser, userInput, expectedCommand);
 
         // address
         userInput = targetIndex.getOneBased() + ADDRESS_DESC_AMY;
         descriptor = new EditPersonDescriptorBuilder().withAddress(VALID_ADDRESS_AMY).build();
-        expectedCommand = new EditCommand(targetIndex, descriptor, EditCommand.ContactType.CLIENT);
+        expectedCommand = new EditCommand(targetIndex, descriptor, Command.ContactType.CLIENT);
         assertParseSuccess(parser, userInput, expectedCommand);
 
         // tags
         userInput = targetIndex.getOneBased() + TAG_DESC_FRIEND;
         descriptor = new EditPersonDescriptorBuilder().withTags(VALID_TAG_FRIEND).build();
-        expectedCommand = new EditCommand(targetIndex, descriptor, EditCommand.ContactType.CLIENT);
+        expectedCommand = new EditCommand(targetIndex, descriptor, Command.ContactType.CLIENT);
         assertParseSuccess(parser, userInput, expectedCommand);
     }
 
@@ -176,7 +177,7 @@ public class EditCommandParserTest {
         EditCommand.EditContactDescriptor descriptor = new EditPersonDescriptorBuilder().withPhone(VALID_PHONE_BOB)
                 .withEmail(VALID_EMAIL_BOB).withAddress(VALID_ADDRESS_BOB).withTags(VALID_TAG_FRIEND, VALID_TAG_HUSBAND)
                 .build();
-        EditCommand expectedCommand = new EditCommand(targetIndex, descriptor, EditCommand.ContactType.CLIENT);
+        EditCommand expectedCommand = new EditCommand(targetIndex, descriptor, Command.ContactType.CLIENT);
 
         assertParseSuccess(parser, userInput, expectedCommand);
     }
@@ -188,7 +189,7 @@ public class EditCommandParserTest {
         String userInput = targetIndex.getOneBased() + INVALID_PHONE_DESC + PHONE_DESC_BOB;
         EditCommand.EditContactDescriptor descriptor = new EditPersonDescriptorBuilder()
                 .withPhone(VALID_PHONE_BOB).build();
-        EditCommand expectedCommand = new EditCommand(targetIndex, descriptor, EditCommand.ContactType.CLIENT);
+        EditCommand expectedCommand = new EditCommand(targetIndex, descriptor, Command.ContactType.CLIENT);
         assertParseSuccess(parser, userInput, expectedCommand);
 
         // other valid values specified
@@ -196,7 +197,7 @@ public class EditCommandParserTest {
                 + PHONE_DESC_BOB;
         descriptor = new EditPersonDescriptorBuilder().withPhone(VALID_PHONE_BOB).withEmail(VALID_EMAIL_BOB)
                 .withAddress(VALID_ADDRESS_BOB).build();
-        expectedCommand = new EditCommand(targetIndex, descriptor, EditCommand.ContactType.CLIENT);
+        expectedCommand = new EditCommand(targetIndex, descriptor, Command.ContactType.CLIENT);
         assertParseSuccess(parser, userInput, expectedCommand);
     }
 
@@ -206,7 +207,7 @@ public class EditCommandParserTest {
         String userInput = targetIndex.getOneBased() + TAG_EMPTY;
 
         EditContactDescriptor descriptor = new EditPersonDescriptorBuilder().withTags().build();
-        EditCommand expectedCommand = new EditCommand(targetIndex, descriptor, EditCommand.ContactType.CLIENT);
+        EditCommand expectedCommand = new EditCommand(targetIndex, descriptor, Command.ContactType.CLIENT);
 
         assertParseSuccess(parser, userInput, expectedCommand);
     }
